doc-qa-ingestor/requirements.txt
fastapi
uvicorn
langchain
boto3
faiss-cpu
numpy
